ai-server/requirements.txt
annotated-types
anyio
blinker
certifi
charset-normalizer
click
community
distro
exceptiongroup
fastapi
Flask
groq
h11
httpcore
httpx
idna
itsdangerous
Jinja2
joblib
MarkupSafe
neo4j
networkx
nltk
pydantic
pydantic_core
python-dotenv
pytz
regex
requests
sniffio
starlette
tqdm
typing_extensions
urllib3
uvicorn
Werkzeug
